refactor: Move resolvable? from Browser to Dns library
